Re-organizing some test stuff. * Bigquery and storage tests extend from common tests.CloudBaseTest. * All test resources now live in tests/resources, accessible via self.resources inside of tests. * Updated tox to allow testing a single file for dev purposes.
