--- bigquery/test/test_load_data_from_csv.py
@@ -1,45 +1,0 @@
-# Copyright 2015, Google, Inc.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-"""Tests for load_data_from_csv."""
-
-import json
-import os
-import unittest
-
-from bigquery.samples.load_data_from_csv import run
-from bigquery.test import RESOURCE_PATH
-from bigquery.test.base_test import BaseBigqueryTest
-
-
-class TestLoadDataFromCSV(BaseBigqueryTest):
-
-    def setUp(self):
-        super(TestLoadDataFromCSV, self).setUp()
-        with open(
-                os.path.join(RESOURCE_PATH, 'schema.json'),
-                'r') as schema_file:
-            self.schema = json.load(schema_file)
-
-    def test_load_table(self):
-        run(self.schema,
-            self.constants['cloudStorageInputURI'],
-            self.constants['projectId'],
-            self.constants['datasetId'],
-            self.constants['newTableId'],
-            5,
-            5)
-
-
-if __name__ == '__main__':
-    unittest.main()
