+++ tox.ini
@@ -22,5 +22,5 @@ commands =
     cover: --cover-tests --cover-branches --cover-min-percentage=70 \
     cover: --cover-inclusive --cover-erase
     cover: coveralls
-    nosetest: nosetests --with-gae
+    nosetest: nosetests --with-gae {posargs}
     pep8: flake8 --max-complexity=10 --import-order-style=google
